class moved to main.rb
I've moved the class to the main.rb file due to all the method from there uses puts. and gets.chomp
