+++ bin/main.rb
@@ -1,6 +1,5 @@
 #!/usr/bin/env ruby
 require_relative '../lib/board.rb'
-require_relative '../lib/player.rb'
 
 def updated_board(board_array)
   puts "#{board_array[0]} | #{board_array[1]} | #{board_array[2]}"
@@ -10,34 +9,47 @@ def updated_board(board_array)
   puts "#{board_array[6]} | #{board_array[7]} | #{board_array[8]}"
 end
 
-def win_or_draw(accu, player, combination_array)
-  win = false
-  combination_array.each { |elem| elem.all?('X') || elem.all?('O') ? win = true : false }
-  if win
-    puts "Player #{player.name} WINS"
-    false
-  elsif accu == 9
-    puts 'No more plays available. Its a DRAW'
-    false
-  else
-    true
-  end
-end
+class Player
+  attr_reader :name
 
-def select_slot(band, turn, board)
-  puts 'Please select one empty space'
-  number = gets.chomp.to_i
-  if (1..9).include?(number)
-    if board.board_array[number - 1] == 'X' || board.board_array[number - 1] == 'O'
-      puts 'Slot occupied'
-    else
-      board.modify_board(number, turn)
-      band = true
-    end
-  else
-    puts 'No such position'
+  def initialize(name)
+    @name = name
   end
-  band
+
+  def player_turn(board, turn)
+    band = false
+    band = select_slot(band, turn, board) until band
+  end
+
+  def select_slot(band, turn, board)
+    puts 'Please select one empty space'
+    number = gets.chomp.to_i
+    if (1..9).include?(number)
+      if board.board_array[number - 1] == 'X' || board.board_array[number - 1] == 'O'
+        puts 'Slot occupied'
+      else
+        board.modify_board(number, turn)
+        band = true
+      end
+    else
+      puts 'No such position'
+    end
+    band
+  end
+
+  def win_or_draw(accu, player, combination_array)
+    win = false
+    combination_array.each { |elem| elem.all?('X') || elem.all?('O') ? win = true : false }
+    if win
+      puts "Player #{player.name} WINS"
+      false
+    elsif accu == 9
+      puts 'No more plays available. Its a DRAW'
+      false
+    else
+      true
+    end
+  end
 end
 
 s_array = %w[S s]
@@ -80,7 +92,7 @@ if p_array.include? key
       updated_board(board.board_array)
       player.player_turn(board, turn)
       accu += 1
-      game_on = win_or_draw(accu, player, board.combination_array)
+      game_on = player.win_or_draw(accu, player, board.combination_array)
       turn = !turn
     end
   elsif player_coin != coin
@@ -90,7 +102,7 @@ if p_array.include? key
       updated_board(board.board_array)
       player.player_turn(board, turn)
       accu += 1
-      game_on = win_or_draw(accu, player, board.combination_array)
+      game_on = player.win_or_draw(accu, player, board.combination_array)
       turn = !turn
     end
   end
